backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pgvector
rq
redis
google-cloud-storage
google-cloud-run
openai
openrouter
python-multipart
python-dotenv
pydantic>=2.0.0
imageio-ffmpeg>=0.4.9

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
